+++ src/redturtle/prenotazioni/behaviors/booking_folder/notifications/email/__init__.py
@@ -299,6 +299,15 @@ class INotificationEmail(model.Schema):
         defaultFactory=notify_as_reminder_message_default_factory,
         required=False,
     )
+    attach_qrcode = schema.Bool(
+        title=_("attach_qrcode_label", default="Attach QR Code"),
+        description=_(
+            "attach_qrcode_help",
+            default="Attach a QR Code with booking code to the email notification.",
+        ),
+        default=False,
+        required=False,
+    )
 
     model.fieldset(
         "email_notifications",
@@ -325,6 +334,7 @@ class INotificationEmail(model.Schema):
             "notify_on_cancel_message",
             "notify_as_reminder_subject",
             "notify_as_reminder_message",
+            "attach_qrcode",
         ],
     )